Add job for compile sass
